requirements.txt
##python version==3.11.3
google-generativeai==0.5.2
pymongo==4.8.0
flask==3.0.3
python-dotenv
pillow
# langchain-google-genai==1.0.8
# langchain==0.2.14
llama-index-llms-gemini==0.1.9
llama-index
llama-index-multi-modal-llms-gemini==0.1.8
streamlit==1.38.0